DFCxx: bitrange boundaries have uint16_t types now.

+++ src/model/dfcxx/include/dfcxx/node.h
@@ -47,8 +47,8 @@ union NodeData {
   uint64_t muxId;
   uint8_t bitShift;
   struct {
-    uint8_t left;
-    uint8_t right;
+    uint16_t left;
+    uint16_t right;
   } bitsRange;
 };
